+++ src/plugin.cpp
@@ -6,6 +6,9 @@ void init(Plugin* p) {
 	pluginInstance = p;
 
 	// Add modules here
+	p->addModel(modelBlank901);
+	p->addModel(modelBlank902);
+	p->addModel(modelBlank904);
 	p->addModel(modelVBNO);
 
 	// Any other plugin initialization may go here.