+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.0.2 (5/19/2020)
 - Added custom knobs & inputs
+- Added Blank901, Blank902, Blank904 Blank Panels
 
 ### 1.0.1 (5/18/2020)
 - Fixed VBNO - Wave input CV - 10v